% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorsplitteR.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{volcanoPlot function by Daniel Crouch
Produces volcano plots with points coloured by priorityFDRs and FDRs, using output from priorityFDR.priorsplitteR
See Crouch et al. 2022, https://doi.org/10.1101/2021.02.05.429962, for details.}
\usage{
volcanoPlot(beta, s, label, priorityFDR.out, FDRthresh = 0.01,
  priorityFDRthresh = 0.01, priorityFDRmeasure = "tailpriorityFDR",
  FDRmeasure = "tailFDR")
}
\arguments{
\item{beta}{Vector of effect size estimates with normal sampling error. Must be supplied}

\item{s}{Vector of standard error estimates, assumed to be close to their true values. Must be supplied}

\item{seed}{Seed for random number generation. Must be supplied}
}
\description{
volcanoPlot function by Daniel Crouch
Produces volcano plots with points coloured by priorityFDRs and FDRs, using output from priorityFDR.priorsplitteR
See Crouch et al. 2022, https://doi.org/10.1101/2021.02.05.429962, for details.
}
