% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorsplitteR.R
\name{priorityFDR.priorsplitteR}
\alias{priorityFDR.priorsplitteR}
\title{priorityFDR.priorsplitteR() function by Daniel Crouch
Estimates priorityFDRs from a set of effect size estimates and standard errors. Effect size estimates are assumed to have normal sampling error.
Returns local and tail area priorityFDRs, effect priorities and FDRs. 
See Crouch et al. 2022, https://doi.org/10.1101/2021.02.05.429962, for details.}
\usage{
priorityFDR.priorsplitteR(beta, s, seed, excludeFromFit = NULL,
  simulationTruth = NULL, outFileStem, breaks = 50,
  degree_altNull = 10, degree_doublePolynomial = 6, nSim = 5e+06,
  diffSampleSize = 500, diffSampleSizeHighAcc = 4500,
  highAccThresh = 0.02, blockSize = NULL, EMtol = c(1e-08, 0.001),
  nStop = 5, maxEMit = c(1e+05, 10000), ml.reltol = 1e-300,
  outlierSD = 10, diffOutlierSD = NULL, truncatedNull = NULL,
  nullTol = 1e-300, threads = 1, modelOnly = FALSE,
  performOneTail = FALSE)
}
\arguments{
\item{beta}{Vector of effect size estimates with normal sampling error. Must be supplied}

\item{s}{Vector of standard error estimates, assumed to be close to their true values. Must be supplied}

\item{seed}{Seed for random number generation. Must be supplied}

\item{excludeFromFit}{Vector of integers specifying the variables to exclude from distribution modelling}

\item{simulationTruth}{Vector of true effect sizes if known from a simulation, for plotting}

\item{outFileStem}{File name stem for output files. Must be supplied}

\item{breaks}{Number of breaks in histogram for Poisson regression distribution modelling, default 50}

\item{degree_altNull}{Degree of polynomial model for ANE model, default 10}

\item{degree_doublePolynomial}{Degree of polynomial-like model for NUPE and NUPE-diff models, default 6}

\item{nSim}{Number of simulated z-scores for NUPE and NUPE-diff models, default 5 million}

\item{diffSampleSize}{Number of 'k' variables to sample for estimating pairwise posterior probabilities for SNP k exceeding focal SNP j in effect size, default 500}

\item{diffSampleSizeHighAcc}{Number of additional 'k' variables to sample for estimating pairwise posterior probabilities for SNP k exceeding focal SNP j in effect size, when estimated local effectPriority_j<highAccThresh, default 4500 giving 4500+500=5000 total samples}

\item{highAccThresh}{Threshold for re-sampling variables to improve estimates of priorityfdr_j based on whether it is low enough to be potentially interesting, default 0.02}

\item{blockSize}{Number of 'j' variables to vectorise priorityfdr_j for, default 1000. When more memory is available, blockSize can be increased with potential improvements in speed}

\item{EMtol}{Vector of length 2 containing tolerances for likelihood convergence in EM algorithm (ANE and NUPE/NUPE-diff respectively), default c(1e-8,1e-3)}

\item{nStop}{Number of consecutive EM iterations that must satisfy EM likelihood convergence tolerance before convergence is accepted, default 5}

\item{maxEMit}{Vector of length 2 containing maximum number of EM iterations before convergence is accepted, for ANE and NUPE/NUPE-diff respectively, default c(100000,10000). Estimates obtained when the maximum number of iterations are reached should be treated with caution}

\item{ml.reltol}{Maximum likelihood tolerance, passed to constrOptim, default 1e-300}

\item{outlierSD}{Variables with absolute z-scores greater than this threshold are removed from distribution modelling. For producing priorityFDR estimates, they are squashed to have absolute values no greater than the maximum value used for modelling. Default 15}

\item{diffOutlierSD}{Pairwise effect estimate differences with absolute z-scores greater than this threshold are removed from distribution modelling. For producing priorityFDR estimates, they are squashed to have absolute values no greater than the maximum value used for modelling. Default NULL, which sets the threshold to outlierSD*1.5}

\item{truncatedNull}{Range within which observed estimates are known a priori to exist, a vector of length 2. For example, if it is known that estimates can only be positive, specify c(0,Inf). Default NULL for range truncation.}

\item{nullTol}{Values with null densities equal to or below this value are omitted from fitting and assumed to have either positive or negative true effects (for positive and negative estimates respectively) with posterior probability 1, and similarly for effect differences. The estimation prodecure fits a model as a ratio against the null distribution, so when null density is recorded as zero no meaningful fit can be produced, and when it is close to zero may be unreliable.}

\item{threads}{Number of threads for parallelising Bayesian computations for pairwise variable effect size comparisons, default 1. We recommended higher settings when multiple threads are available}

\item{modelOnly}{If TRUE, only model estimation is performed, returning estimated parameters. PriorityFDRs, FDRs and effect priorities are not computed for each variant}

\item{performOneTail}{If true, one-tail inclusive priorityFDRs are computed, estimating the probability that a variable's true effect, rather than effect size, is exceeded by a randomly chosen variant. This requires additional computational time but the results can be used to estimate the prior distribution of effects using priorEst.priorsplitteR}
}
\description{
priorityFDR.priorsplitteR() function by Daniel Crouch
Estimates priorityFDRs from a set of effect size estimates and standard errors. Effect size estimates are assumed to have normal sampling error.
Returns local and tail area priorityFDRs, effect priorities and FDRs. 
See Crouch et al. 2022, https://doi.org/10.1101/2021.02.05.429962, for details.
}
